refactor: extract audio analysis normalizers to separate module
